Improved the download script (#3)
* Improved script to download quadlet-lsp artifacts

Improvements to portability and error-resilience
overall. Reduced code duplication using a loop

* Removed build artifacts from git repository:

Since the artifacts can be downloaded, it doesn't
make sense to include them in the git repository.

The .gitignore is updated to reflect this

* Included darwin-arm64 in the quadlet-lsp download

It was oddly missing, but its a target of both
projects, so I added it.